- Add to allow going back and forth on photo related tasks - Add to allow photo conv to remove photo
