+++ pdf_bot/commands/__init__.py
@@ -1,3 +1,4 @@
 from pdf_bot.commands.compare import compare_cov_handler
 from pdf_bot.commands.merge import merge_cov_handler
 from pdf_bot.commands.watermark import watermark_cov_handler
+from pdf_bot.commands.photo import photo_cov_handler, process_photo
