+++ pdf_bot/constants.py
@@ -44,10 +44,10 @@ ROTATE = _('Rotate')
 SCALE = _('Scale')
 SPLIT = _('Split')
 BEAUTIFY = _('Beautify')
-CONVERT = _('Convert')
+TO_PDF = _('To PDF')
 RENAME = _('Rename')
 CROP = _('Crop')
-ZIPPED = _('Zipped')
+COMPRESSED = _('Compressed')
 PHOTOS = _('Photos')
 REMOVE_LAST = _('Remove Last File')
 TO_DIMENSIONS = _('To Dimensions')
