--- pdf_bot/photos/__init__.py
@@ -1,4 +1,0 @@
-from pdf_bot.photos.pdf_to_photo import get_pdf_preview, get_pdf_photos, pdf_to_photos, \
-    ask_photo_results_type
-from pdf_bot.photos.photo_to_pdf import photo_cov_handler, process_photo, process_photo_task, \
-    ask_photo_task
